requirements.txt
langgraph
langchain
langchain-openai
langchain-community
langchain-text-splitters
fastapi
uvicorn
pydantic
pydantic-settings
qdrant-client
sentence-transformers
openai
python-dotenv
numpy<2.0
pandas
scikit-learn
nltk
spacy
textstat
rouge-score
bert-score
transformers
torch
typing-extensions
pytest
pytest-asyncio
aiofiles
httpx
requests
aiohttp
